- Added Linearity for 256x265 resolution in PAL and NTSC - Added help and details for Alternate 240p/480i test - Moved video options to a general menu - Cleaned up the code formating

+++ trunk/240psuite/Genesis/240p/myvdp.h
@@ -30,8 +30,8 @@
 
 void VDP_setSpriteAttr(u16 index, u16 tile_attr);
 void VDP_setMyTileMapRect(u16 plan, const u16 *data, u16 basetile, u16 x, u16 y, u16 w, u16 h);
-u16  Detect_VDP_PAL();
-u16 VDP_Detect_Interlace();
+u16 Detect_VDP_PAL();
+u8 VDP_Detect_Interlace();
 
 extern u16 pal_240;
